rename, fix logic
